step4/pyproject.toml
[project]
name = "step4"
version = "0.1.0"
description = ""
authors = [
    {name = "Roberto Filho",email = "zferraz.rf@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "colorama (>=0.4.6,<0.5.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "flask-cors (>=5.0.1,<6.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
